refactor(rule): reorganized the code of the topology component
